+++ epam-electric/components/widgets/StatusIndicator.tsx
@@ -1,13 +1,14 @@
 import { createSkinComponent } from '@epam/uui-core';
 import * as uui from '@epam/uui';
 
-export type StatusIndicatorMods = {
-    /** Represents the color of a status indicator.
+type StatusIndicatorMods = {
+    /** Defines component color.
      * @default 'gray'
      */
     color?: uui.StatusIndicatorColors | 'white' | 'gray' | 'yellow'| 'orange' | 'fuchsia' | 'purple' | 'violet' | 'cobalt' | 'cyan' | 'mint';
 };
 
+/** Represents the properties of a StatusIndicator component. */
 export type StatusIndicatorProps = uui.StatusIndicatorCoreProps & StatusIndicatorMods;
 
 export const StatusIndicator = createSkinComponent<uui.StatusIndicatorProps, StatusIndicatorProps>(