+++ epam-electric/components/widgets/Tag.tsx
@@ -3,13 +3,13 @@ import * as uui from '@epam/uui';
 
 export type TagMods = {
     /**
-     * The color property of a Tag component.
+     * Defines component color.
      * @default 'neutral'
      */
     color?: uui.TagColor | 'white' | 'night100' | 'night700';
 };
 
-/** Represents the properties of a tag component. */
+/** Represents the properties of a Tag component. */
 export type TagProps = uui.TagCoreProps & TagMods;
 
 export const Tag = createSkinComponent<uui.TagProps, TagProps>(uui.Tag);