+++ loveship/components/inputs/index.ts
@@ -1,5 +1,4 @@
 export * from './Slider';
 export * from './Rating';
-export * from './RadioInput';
 export * from './MultiSwitch';
 export * from './SliderRating';