+++ loveship/components/widgets/Badge.tsx
@@ -6,30 +6,30 @@ import css from './Badge.module.scss';
 
 const defaultSize = '18';
 
-export type BadgeMods = {
+type BadgeMods = {
     /**
-     * The color variable represents the color of an object.
+     * TDefines component color.
      * @default 'sky'
      */
     color?: EpamPrimaryColor | EpamAdditionalColor | 'yellow'| 'orange' | 'purple' | 'cyan' | 'mint' | 'white' | 'night100' | 'night300' | 'night600';
     /**
-     * Represents the shape of a control.
+     * Defines component shape.
      * @default 'square'
      */
     shape?: types.ControlShape;
     /**
-     * The fill property for the BadgeMods object.
+     * Defines component fill style.
      * @default 'solid'
      */
     fill?: uui.BadgeMods['fill'] | 'semitransparent';
     /**
-     * The size of the badge.
+     * Defines component size.
      * @default '18'
      */
     size?: uui.BadgeMods['size'] | '12';
 };
 
-export function applyBadgeMods(mods: BadgeMods) {
+function applyBadgeMods(mods: BadgeMods) {
     return [
         css['style-' + (mods.shape || 'square')],
         css[`fill-${mods.fill === 'semitransparent' ? 'outline' : (mods.fill || 'solid')}`],
@@ -38,9 +38,7 @@ export function applyBadgeMods(mods: BadgeMods) {
     ];
 }
 
-/**
- * Represents the props for the Badge component.
- */
+/** Represents the properties of a Badge component. */
 export type BadgeProps = uui.BadgeCoreProps & BadgeMods;
 
 export const Badge = createSkinComponent<uui.BadgeProps, BadgeProps>(