+++ uui/components/inputs/TextInput.tsx
@@ -5,24 +5,18 @@ import { IHasEditMode, EditMode, ControlSize } from '../types';
 import { systemIcons } from '../../icons/icons';
 import css from './TextInput.module.scss';
 
-/**
- * Represents the default size for a specific feature.
- */
 const defaultSize = '36';
-/**
- * Represents the default mode for editing a form.
- */
 const defaultMode = EditMode.FORM;
 
-export interface TextInputMods extends IHasEditMode {
+type TextInputMods = IHasEditMode & {
     /**
-     * Specifies the size of the control.
+     * Defines component size.
      * @default '36'
      */
     size?: ControlSize | '60';
-}
+};
 
-export function applyTextInputMods(mods: TextInputMods) {
+function applyTextInputMods(mods: TextInputMods) {
     return [
         css.root,
         css['size-' + (mods.size || defaultSize)],
@@ -30,13 +24,10 @@ export function applyTextInputMods(mods: TextInputMods) {
     ];
 }
 
-/**
- * Represents the properties for a text input component.
- */
+/** Represents the properties for a TextInput component. */
 export interface TextInputProps extends CoreTextInputProps, TextInputMods {}
-/**
- * Represents the properties for the SearchInput component.
- */
+
+/** Represents the properties for a SearchInput component. */
 export interface SearchInputProps extends TextInputProps, IEditableDebouncerOptions {}
 
 export const TextInput = withMods<CoreTextInputProps, TextInputMods>(uuiTextInput, applyTextInputMods, (props) => ({